convert all metadata.tsv files in package to long form

This function runs through the metadata.tsv files in the installed package, running `study_long_form()` on each, then binds them together, appending a `study` column.

metadata_long_form()

Value

a data.frame with the following columns:

- study - sampleID - subjectID - variable: the column name from the wide-form data - numeric_value: if a numeric value, this value will be non-NA - string_value: if a character value, this value will be non-NA - logical_value: if a logical value, this value will be non-NA

Author

Sean Davis <seandavi@gmail.com>

Examples

ret = metadata_long_form()
dim(ret)

#> [1] 463979      7

ret

#> # A tibble: 463,979 x 7
#>    study  sampleID  subjectID variable  numeric_value logical_value string_value
#>    <chr>  <chr>     <chr>     <chr>             <dbl> <lgl>         <chr>       
#>  1 Asnic… MV_FEI1_… MV_FEI1   age                   1 NA            NA          
#>  2 Asnic… MV_FEI1_… MV_FEI1   infant_a…            90 NA            NA          
#>  3 Asnic… MV_FEI1_… MV_FEI1   PMID           28144631 NA            NA          
#>  4 Asnic… MV_FEI1_… MV_FEI1   number_r…      27553455 NA            NA          
#>  5 Asnic… MV_FEI1_… MV_FEI1   number_b…    2748531713 NA            NA          
#>  6 Asnic… MV_FEI1_… MV_FEI1   minimum_…            80 NA            NA          
#>  7 Asnic… MV_FEI1_… MV_FEI1   median_r…           100 NA            NA          
#>  8 Asnic… MV_FEI2_… MV_FEI2   age                   1 NA            NA          
#>  9 Asnic… MV_FEI2_… MV_FEI2   infant_a…            90 NA            NA          
#> 10 Asnic… MV_FEI2_… MV_FEI2   PMID           28144631 NA            NA          
#> # … with 463,969 more rows

table(ret$study)

#> 
#>          AsnicarF_2017          AsnicarF_2021          BackhedF_2015 
#>                    504                  20862                   9600 
#>  Bengtsson-PalmeJ_2015           BritoIL_2016           BrooksB_2017 
#>                   1400                   6240                  10200 
#>    Castro-NallarE_2015        ChengpingW_2017            ChngKR_2016 
#>                    608                   3007                   1638 
#>             ChuDM_2017      ContevilleLC_2019          CosteaPI_2017 
#>                   3060                    300                   5580 
#>           DavidLA_2015       DeFilippisF_2019          DhakanDB_2019 
#>                   1081                   1843                   2310 
#>             FengQ_2015         FerrettiP_2018           GhensiP_2019 
#>                   3850                   4730                   2825 
#>   GopalakrishnanV_2018            GuptaA_2019            HallAB_2017 
#>                    375                   1500                   5957 
#>        HanniganGD_2017         HansenLBS_2018   Heitz-BuschartA_2016 
#>                   1722                   4347                   1325 
#>               HMP_2012        HMP_2019_ibdmdb           HMP_2019_t2d 
#>                  14980                  37674                  12177 
#>            IjazUZ_2017              JieZ_2017        KarlssonFH_2013 
#>                   2350                  11165                   5220 
#>             KaurK_2020         KeohaneDM_2020           KieserS_2018 
#>                    806                   2691                    486 
#>          KosticAD_2015         LassalleF_2017      LeChatelierE_2013 
#>                   3240                    456                   5256 
#>     LifeLinesDeep_2016               LiJ_2014               LiJ_2017 
#>                  22700                   6240                   3724 
#>              LiSS_2016              LiuW_2016           LokmerA_2019 
#>                    990                   1980                   1197 
#>           LomanNJ_2013           LoombaR_2017            LouisS_2016 
#>                    903                   1634                   1656 
#>           MatsonV_2018           MehtaRS_2018       NagySzakalD_2017 
#>                    663                  16704                   1900 
#>         NielsenHB_2014    Obregon-TitoAJ_2015               OhJ_2014 
#>                   8712                   1160                   5238 
#>             OlmMR_2017          PasolliE_2019         PehrssonE_2016 
#>                    945                   2464                   3820 
#>              QinJ_2012              QinN_2014         RampelliS_2015 
#>                   7986                   6399                    722 
#>          RaymondF_2016            RosaBA_2018           RubelMA_2020 
#>                   1728                    504                   3850 
#> SankaranarayananK_2015         SchirmerM_2016             ShaoY_2019 
#>                    777                  11304                  46032 
#>              ShiB_2015           SmitsSA_2017            TettAJ_2016 
#>                   1008                    720                   1940 
#>          TettAJ_2019_a          TettAJ_2019_b          TettAJ_2019_c 
#>                   1360                    880                   1000 
#>         ThomasAM_2018a         ThomasAM_2018b        ThomasAM_2019_c 
#>                   1840                   1200                   1520 
#>          VatanenT_2016            VilaAV_2018          VincentC_2016 
#>                  19625                   7100                   4580 
#>         VogtmannE_2016          WampachL_2018              WenC_2017 
#>                   2090                   1386                   5544 
#>           WirbelJ_2018              XieH_2016          YachidaS_2019 
#>                   2625                   8000                  12320 
#>          YassourM_2016          YassourM_2018               YeZ_2018 
#>                    864                   7588                   1300 
#>               YuJ_2015            ZeeviD_2015           ZellerG_2014 
#>                   3200                  15300                   3588 
#>              ZhuF_2020 
#>                   4104 

summary(ret)

#>     study             sampleID          subjectID           variable        
#>  Length:463979      Length:463979      Length:463979      Length:463979     
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>  numeric_value       logical_value  string_value      
#>  Min.   :0.000e+00   Mode:logical   Length:463979     
#>  1st Qu.:4.000e+01   NA's:463979    Class :character  
#>  Median :1.010e+02                  Mode  :character  
#>  Mean   :4.856e+08                                    
#>  3rd Qu.:2.865e+07                                    
#>  Max.   :6.973e+10                                    
#>  NA's   :297644